FIX: Textdraws not showing, City selection bug
Instead of using returns to stop functions, I for some reason tried using "pass", causing my statements to just carry on and not stop the code.

Textdraws were not shown because the variables were not assigned in to the local scopes using "global" keywords, hence the vars were empty.

Also updated the samp.py according to the latest update, that implemented 3D Textlabels