FIX: Spawn issue with TogglePlayerSpectate()
The trick was to spawn the player first to make everything load, and then toggle on the spectation. Else the code freaks out and disconnects the client.  

Also added the newest update in the samp.py file into the grand larceny gamemode folder.
Some comments added to inform the intention behind this rewrite and also re-enabled some checks for illegal weapons and actions under onplayerupdate. Tested and ready for release.