Add Grand Larceny to gamemodes (#44)
* WIP: Grand Larceny rewrite to PySAMP

Work still in progress, and nothing has been tested yet. Not ready for use.

* Finished translation to python

Needs further testing to see if functionality is same as on pawn, also find the base bath for opening files.

* Fix: type assignments, vehicle loading

Fixed a few integers where they were pushed as strings. Gamemode loads without errors at this point, in-game testing starting

* WIP: Player Spawning not working correctly

I'm having issues after line 317 kicks in. OnPlayerSpawn is not called and the client is informed "Server closed the connection". 

Work in progress figuring that out.

FIX: Removing player objects when leaving

* FIX: Textdraws not showing, City selection bug

Instead of using returns to stop functions, I for some reason tried using "pass", causing my statements to just carry on and not stop the code.

Textdraws were not shown because the variables were not assigned in to the local scopes using "global" keywords, hence the vars were empty.

Also updated the samp.py according to the latest update, that implemented 3D Textlabels

* FIX: Array entries were wrong

* FIX: Spawn issue with TogglePlayerSpectate()

The trick was to spawn the player first to make everything load, and then toggle on the spectation. Else the code freaks out and disconnects the client.  

Also added the newest update in the samp.py file into the grand larceny gamemode folder.
Some comments added to inform the intention behind this rewrite and also re-enabled some checks for illegal weapons and actions under onplayerupdate. Tested and ready for release.

Co-authored-by: denNorske <den@ducky.rocks>